fix(engine): harden hash stability and security across core modules
Prevent hash collision vulnerabilities by length-prefixing all variable-length
fields before hashing, and fix several security/correctness issues:

- cache_oblivious_metadata_substrate: use stable byte discriminants instead of
  Debug format for hash inputs, preventing hash instability across Rust versions
- expected_loss_selector: length-prefix action/state strings and prepend entry
  count to prevent concatenation collisions in loss matrix hashing
- proof_ingestion: replace naive H(key||msg) signature construction with proper
  AuthenticityHash keyed construction to prevent length-extension attacks; add
  #[serde(skip)] to signing_key field to prevent key leakage in serialization
- recovery_artifact: length-prefix proof and operator action fields in both
  ArtifactBuilder and RecoveryArtifactStore ID computation
- regret_bounded_router: widen Exp3 importance-weighted reward and log-weight
  delta computations to i128 to prevent integer overflow on extreme values
- rough_path_regime_geometry: length-prefix trace_id, signature_hash, and
  corridor from/to strings in topology rehashing
- semantic_flattening_inventory: length-prefix all variable-length fields and
  add explicit Option markers for line_hint presence/absence
- signature_drift_gate: length-prefix signature_id and feature map keys
- succinct_witness_compiler: introduce length-prefix helper closure for
  WitnessSchema ID and ProvenanceAttachment content hash computation; add
  epoch field and explicit Option marker for legal_summary

+++ crates/franken-engine/src/regret_bounded_router.rs
@@ -261,11 +261,13 @@ impl Exp3State {
         let probs = self.arm_probabilities();
         let p_arm = probs[arm].max(1); // avoid division by zero
 
-        // Importance-weighted reward: r̂ = r / p
-        let importance_weighted = reward_millionths * MILLION / p_arm;
+        // Importance-weighted reward: r̂ = r / p (widen to i128 to avoid overflow)
+        let importance_weighted =
+            (reward_millionths as i128 * MILLION as i128 / p_arm as i128) as i64;
 
-        // Update log-weight: log_w += η · r̂ / MILLION
-        let delta = self.eta_millionths * importance_weighted / MILLION;
+        // Update log-weight: log_w += η · r̂ / MILLION (widen to avoid overflow)
+        let delta =
+            (self.eta_millionths as i128 * importance_weighted as i128 / MILLION as i128) as i64;
         self.log_weights_millionths[arm] += delta;
 
         self.rounds += 1;
